Replace absolute positional embeddings with RoPE
